changed xtb/orca modules to always output a new Molecule object instead of updating the input one
