+++ .gitignore
@@ -1,3 +1,5 @@
+main.py
+
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
