+++ compechem/modules/orca.py
@@ -1,5 +1,6 @@
 import os, shutil
 from tempfile import mkdtemp
+from compechem.molecule import Molecule
 
 
 class OrcaInput:
@@ -47,36 +48,40 @@ class OrcaInput:
         self.solvent = solvent
         self.optionals = optionals
 
-    def spe(self, mol, remove_tdir=True):
+    def spe(self, mol, charge=None, spin=None, remove_tdir=True):
         """Single point energy calculation.
 
         Parameters
         ----------
         mol : Molecule object
             input molecule to use in the calculation
+        charge : int, optional
+            total charge of the molecule. Default is taken from the input molecule.
+        spin : int, optional
+            total spin of the molecule. Default is taken from the input molecule.
         remove_tdir : bool, optional
             temporary work directory will be removed, by default True
 
         Returns
         -------
-        Updates the mol.energies dictionary adding an entry with:
-            method
-            electronic energy
+        newmol : Molecule object
+            Output molecule containing the new energies.
         """
 
+        if charge is None:
+            charge = mol.charge
+        if spin is None:
+            spin = mol.spin
+
         parent_dir = os.getcwd()
-        print(
-            f"INFO: {mol.name}, charge {mol.charge} spin {mol.spin} - {self.method} SPE"
-        )
+        print(f"INFO: {mol.name}, charge {charge} spin {spin} - {self.method} SPE")
 
         tdir = mkdtemp(
-            prefix=mol.name + "_",
-            suffix=f"_{self.method.split()[0]}_spe",
-            dir=os.getcwd(),
+            prefix=mol.name + "_", suffix=f"_{self.method.split()[0]}_spe", dir=os.getcwd(),
         )
 
         os.chdir(tdir)
-        mol.write_xyz("geom.xyz")
+        mol.write_xyz(f"{mol.name}.xyz")
 
         with open("input.inp", "w") as inp:
             inp.write(
@@ -86,22 +91,21 @@ class OrcaInput:
                 f"! RIJCOSX {self.aux_basis}\n"
             )
             if self.solvation is True:
-                inp.write(
-                    "%CPCM\n"
-                    "  SMD True\n"
-                    f'  SMDsolvent "{self.solvent}"\n'
-                    "end\n"
-                )
-            inp.write(f"* xyzfile {mol.charge} {mol.spin} geom.xyz\n")
+                inp.write("%CPCM\n" "  SMD True\n" f'  SMDsolvent "{self.solvent}"\n' "end\n")
+            inp.write(f"* xyzfile {charge} {spin} {mol.name}.xyz\n")
 
         os.system("$ORCADIR/orca input.inp > output.out")
+
+        newmol = Molecule(f"{mol.name}.xyz", charge, spin)
+
+        newmol.energies = mol.energies
 
         with open("output.out", "r") as out:
             for line in out:
                 if "FINAL SINGLE POINT ENERGY" in line:
                     electronic_energy = float(line.split()[-1])
 
-        mol.energies[f"{self.method}"] = mol.Energies(
+        newmol.energies[f"{self.method}"] = newmol.Energies(
             method=f"{self.method}", electronic=electronic_energy,
         )
 
@@ -109,108 +113,61 @@ class OrcaInput:
             shutil.rmtree(tdir)
         os.chdir(parent_dir)
 
-        return
+        return newmol
 
-    def opt(self, mol, remove_tdir=True):
+    def opt(self, mol, charge=None, spin=None, remove_tdir=True):
         """Geometry optimization + frequency analysis.
 
         Parameters
         ----------
         mol : Molecule object
             input molecule to use in the calculation
+        charge : int, optional
+            total charge of the molecule. Default is taken from the input molecule.
+        spin : int, optional
+            total spin of the molecule. Default is taken from the input molecule.
         remove_tdir : bool, optional
             temporary work directory will be removed, by default True
 
         Returns
         -------
-        Updates the molecular geometry
+        newmol : Molecule object
+            Output molecule containing the new geometry and energies.
         """
 
+        if charge is None:
+            charge = mol.charge
+        if spin is None:
+            spin = mol.spin
+
         parent_dir = os.getcwd()
-        print(
-            f"INFO: {mol.name}, charge {mol.charge} spin {mol.spin} - {self.method} OPT"
-        )
+        print(f"INFO: {mol.name}, charge {charge} spin {spin} - {self.method} OPT")
 
         tdir = mkdtemp(
-            prefix=mol.name + "_",
-            suffix=f"_{self.method.split()[0]}_opt",
-            dir=os.getcwd(),
+            prefix=mol.name + "_", suffix=f"_{self.method.split()[0]}_opt", dir=os.getcwd(),
         )
 
         os.chdir(tdir)
-        mol.write_xyz("geom.xyz")
+        mol.write_xyz(f"{mol.name}.xyz")
 
         with open("input.inp", "w") as inp:
             inp.write(
                 f"%pal nproc {self.nproc} end\n"
                 f"%maxcore {self.maxcore}\n"
                 f"! {self.method} {self.basis_set} {self.optionals}\n"
-                f"! RIJCOSX {self.aux_basis}\n"
+                f"! RIJCOSX {self.aux_basis}"
             )
             if self.solvation is True:
                 inp.write(
+                    "! Opt NumFreq\n"
                     "%CPCM\n"
                     "  SMD True\n"
                     f'  SMDsolvent "{self.solvent}"\n'
                     "end\n"
                 )
-            inp.write(f"* xyzfile {mol.charge} {mol.spin} geom.xyz\n")
-
-        os.system("$ORCADIR/orca input.inp > output.out")
-
-        if remove_tdir is True:
-            shutil.rmtree(tdir)
-        os.chdir(parent_dir)
-
-        return
-
-    def freq(self, mol, remove_tdir=True):
-        """Frequency analysis (analytical frequencies).
-
-        Parameters
-        ----------
-        mol : Molecule object
-            input molecule to use in the calculation
-        remove_tdir : bool, optional
-            temporary work directory will be removed, by default True
-
-        Returns
-        -------
-        Updates the mol.energies dictionary adding an entry with:
-            method
-            electronic energy
-            vibronic contribution
-        """
-
-        parent_dir = os.getcwd()
-        print(
-            f"INFO: {mol.name}, charge {mol.charge} spin {mol.spin} - {self.method} FREQ"
-        )
-
-        tdir = mkdtemp(
-            prefix=mol.name + "_",
-            suffix=f"_{self.method.split()[0]}_freq",
-            dir=os.getcwd(),
-        )
-
-        os.chdir(tdir)
-        mol.write_xyz("geom.xyz")
-
-        with open("input.inp", "w") as inp:
-            inp.write(
-                f"%pal nproc {self.nproc} end\n"
-                f"%maxcore {self.maxcore}\n"
-                f"! {self.method} {self.basis_set} {self.optionals}\n"
-                f"! RIJCOSX {self.aux_basis}\n"
-            )
-            if self.solvation is True:
-                inp.write(
-                    "%CPCM\n"
-                    "  SMD True\n"
-                    f'  SMDsolvent "{self.solvent}"\n'
-                    "end\n"
-                )
-            inp.write(f"* xyzfile {mol.charge} {mol.spin} geom.xyz\n")
+            else:
+                inp.write("! Opt Freq\n")
+            inp.write(f"* xyzfile {charge} {spin} {mol.name}.xyz\n")
 
         os.system("$ORCADIR/orca input.inp > output.out")
 
@@ -221,49 +178,134 @@ class OrcaInput:
                 if "G-E(el)" in line:
                     vibronic_energy = float(line.split()[-4])
 
-        mol.energies[f"{self.method}"] = mol.Energies(
-            method=f"{self.method}",
-            electronic=electronic_energy,
-            vibronic=vibronic_energy,
+        newmol = Molecule(f"{mol.name}.xyz", charge, spin)
+
+        newmol.energies = mol.energies
+
+        newmol.energies[f"{self.method}"] = newmol.Energies(
+            method=f"{self.method}", electronic=electronic_energy, vibronic=vibronic_energy
         )
 
         if remove_tdir is True:
             shutil.rmtree(tdir)
         os.chdir(parent_dir)
 
-        return
+        return newmol
 
-    def nfreq(self, mol, remove_tdir=True):
+    def freq(self, mol, charge=None, spin=None, remove_tdir=True):
+        """Frequency analysis (analytical frequencies).
+
+        Note: if the SMD solvation model is detected, defaults to numerical frequencies
+        (analytical frequencies are not currently supported)
+
+        Parameters
+        ----------
+        mol : Molecule object
+            input molecule to use in the calculation
+        charge : int, optional
+            total charge of the molecule. Default is taken from the input molecule.
+        spin : int, optional
+            total spin of the molecule. Default is taken from the input molecule.
+        remove_tdir : bool, optional
+            temporary work directory will be removed, by default True
+
+        Returns
+        -------
+        newmol : Molecule object
+            Output molecule containing the new energies.
+        """
+
+        if charge is None:
+            charge = mol.charge
+        if spin is None:
+            spin = mol.spin
+
+        parent_dir = os.getcwd()
+        print(f"INFO: {mol.name}, charge {charge} spin {spin} - {self.method} FREQ")
+
+        tdir = mkdtemp(
+            prefix=mol.name + "_", suffix=f"_{self.method.split()[0]}_freq", dir=os.getcwd(),
+        )
+
+        os.chdir(tdir)
+        mol.write_xyz(f"{mol.name}.xyz")
+
+        with open("input.inp", "w") as inp:
+            inp.write(
+                f"%pal nproc {self.nproc} end\n"
+                f"%maxcore {self.maxcore}\n"
+                f"! {self.method} {self.basis_set} {self.optionals}\n"
+                f"! RIJCOSX {self.aux_basis}"
+            )
+            if self.solvation is True:
+                inp.write(
+                    "! NumFreq\n"
+                    "%CPCM\n"
+                    "  SMD True\n"
+                    f'  SMDsolvent "{self.solvent}"\n'
+                    "end\n"
+                )
+            else:
+                inp.write("! Freq\n")
+            inp.write(f"* xyzfile {charge} {spin} {mol.name}.xyz\n")
+
+        os.system("$ORCADIR/orca input.inp > output.out")
+
+        with open("output.out", "r") as out:
+            for line in out:
+                if "FINAL SINGLE POINT ENERGY" in line:
+                    electronic_energy = float(line.split()[-1])
+                if "G-E(el)" in line:
+                    vibronic_energy = float(line.split()[-4])
+
+        newmol = Molecule(f"{mol.name}.xyz", charge, spin)
+
+        newmol.energies = mol.energies
+
+        newmol.energies[f"{self.method}"] = newmol.Energies(
+            method=f"{self.method}", electronic=electronic_energy, vibronic=vibronic_energy
+        )
+
+        if remove_tdir is True:
+            shutil.rmtree(tdir)
+        os.chdir(parent_dir)
+
+        return newmol
+
+    def nfreq(self, mol, charge=None, spin=None, remove_tdir=True):
         """Frequency analysis (numerical frequencies).
 
         Parameters
         ----------
         mol : Molecule object
             input molecule to use in the calculation
+        charge : int, optional
+            total charge of the molecule. Default is taken from the input molecule.
+        spin : int, optional
+            total spin of the molecule. Default is taken from the input molecule.
         remove_tdir : bool, optional
             temporary work directory will be removed, by default True
 
         Returns
         -------
-        Updates the mol.energies dictionary adding an entry with:
-            method
-            electronic energy
-            vibronic contribution
+        newmol : Molecule object
+            Output molecule containing the new energies.
         """
 
+        if charge is None:
+            charge = mol.charge
+        if spin is None:
+            spin = mol.spin
+
         parent_dir = os.getcwd()
-        print(
-            f"INFO: {mol.name}, charge {mol.charge} spin {mol.spin} - {self.method} NFREQ"
-        )
+        print(f"INFO: {mol.name}, charge {charge} spin {spin} - {self.method} NFREQ")
 
         tdir = mkdtemp(
-            prefix=mol.name + "_",
-            suffix=f"_{self.method.split()[0]}_nfreq",
-            dir=os.getcwd(),
+            prefix=mol.name + "_", suffix=f"_{self.method.split()[0]}_nfreq", dir=os.getcwd(),
         )
 
         os.chdir(tdir)
-        mol.write_xyz("geom.xyz")
+        mol.write_xyz(f"{mol.name}.xyz")
 
         with open("input.inp", "w") as inp:
             inp.write(
@@ -271,15 +313,12 @@ class OrcaInput:
                 f"%maxcore {self.maxcore}\n"
                 f"! {self.method} {self.basis_set} {self.optionals}\n"
                 f"! RIJCOSX {self.aux_basis}\n"
+                "! NumFreq\n"
             )
             if self.solvation is True:
-                inp.write(
-                    "%CPCM\n"
-                    "  SMD True\n"
-                    f'  SMDsolvent "{self.solvent}"\n'
-                    "end\n"
-                )
-            inp.write(f"* xyzfile {mol.charge} {mol.spin} geom.xyz\n")
+                inp.write("%CPCM\n" "  SMD True\n" f'  SMDsolvent "{self.solvent}"\n' "end\n")
+
+            inp.write(f"* xyzfile {charge} {spin} {mol.name}.xyz\n")
 
         os.system("$ORCADIR/orca input.inp > output.out")
 
@@ -290,15 +329,17 @@ class OrcaInput:
                 if "G-E(el)" in line:
                     vibronic_energy = float(line.split()[-4])
 
-        mol.energies[f"{self.method}"] = mol.Energies(
-            method=f"{self.method}",
-            electronic=electronic_energy,
-            vibronic=vibronic_energy,
+        newmol = Molecule(f"{mol.name}.xyz", charge, spin)
+
+        newmol.energies = mol.energies
+
+        newmol.energies[f"{self.method}"] = newmol.Energies(
+            method=f"{self.method}", electronic=electronic_energy, vibronic=vibronic_energy
         )
 
         if remove_tdir is True:
             shutil.rmtree(tdir)
         os.chdir(parent_dir)
 
-        return
+        return newmol
